Add html templates and stuff
