+++ labels.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <link type="text/css" rel="stylesheet" href="styles.css">
+        <title>Entrenamiento de K-Means</title>
+    </head>
+    <header style="background-color: lightblue;height: 5vh; display: flex; align-items: center;">
+        <ul style="margin: 0; list-style-type: none; font-weight: bold; padding-left: 20px;">
+            <li><a href="mainPage" style="text-decoration: none;color: black;">Pagina Principal</a></li>
+        </ul>
+    </header>
+    <body style="margin: 0;">
+        <ul>
+            {{range .}}<li>{{ . }}</li>{{end}}
+        </ul>
+    </body>
+</html>
