+++ main.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <link type="text/css" rel="stylesheet" href="styles.css">
+        <title>Entrenamiento de K-Means</title>
+    </head>
+    <header style="background-color: lightblue;height: 5vh; display: flex; align-items: center;">
+        <ul style="margin: 0; list-style-type: none; font-weight: bold; padding-left: 20px;">
+            <li><a href="mainPage" style="text-decoration: none;color: black;">Pagina Principal</a></li>
+        </ul>
+    </header>
+    <body style="margin: 0;">
+        <div>
+            <div style="display: flex; justify-content: center;">
+                <h1>¡Entrena un modelo de K-Means!</h1>
+            </div>
+            <div>
+                <picture style="display: flex; justify-content: center;"><img src="https://prateekvjoshi.com/wp-content/uploads/2013/06/main.jpg"\></picture>
+            </div>
+            <div style="display: flex; justify-content: center;">
+                <button style="border-radius: 20px; font-size: 24px;"><a href="train" style="text-decoration: none;color: black; font-weight: bold;">¡Empieza a entrenar!</a></button>
+            </div>
+        </div>
+    </body>
+</html>
