+++ scripts/dataset1.txt
@@ -1,0 +1,2 @@
+11886.82, 10002.96, 12734.74, 12324.54, 2121.05, 5985.45, 11972.82, 6441.81, 11994.44, 3852.51, 2231.64, 3453.13, 11630.32, 13008, 3830.51, 5663.63, 8363.07, 12067.8, 2913.37, 7418.87, 10774.14, 5780.01, 12448.56, 12306.98, 14123.47, 13173.54, 6412.73, 10141.31, 14333.37, 3606.52, 10168.93, 8354.01, 6114.57, 9250.94, 14058, 7637.92, 6818.66, 1372.12, 4814.87, 5931.59, 12025.32, 1945.53, 7213.4, 13043.68, 12052.64, 4652.24, 12494.92, 13476.42, 11997.26, 13509.18, 1320.45, 14348.55, 11012.42, 8152.8, 3229.62, 5793.82, 11744.59, 12212.27, 7155.63, 9115.56, 4421.84, 10784.34, 9339.03, 8586.65, 2861.95, 11066.36, 7600.66, 3375.7, 12719.85, 7617.71, 11492.94, 12999.91, 8910.17, 4247.01, 7522.04, 8939.6, 14130.59, 2398.98, 5710.79, 3074.29, 6214.92, 4508.92, 14908.81, 9864.66, 13967.29, 1545.81, 14517.38, 4736.17, 2643.23, 10832.05, 3709.8, 14281.59, 12261.9, 13743.45, 5633.04, 2506.67, 9194.68, 1569.8, 11111.93, 10481.22
+54, 61, 33, 75, 69, 36, 73, 79, 33, 44, 73, 53, 29, 46, 74, 19, 67, 72, 66, 51, 77, 54, 44, 33, 57, 71, 65, 65, 47, 74, 61, 27, 38, 33, 78, 67, 52, 65, 35, 78, 36, 74, 78, 71, 43, 50, 19, 72, 65, 39, 48, 35, 21, 54, 61, 27, 50, 49, 25, 48, 71, 51, 52, 52, 76, 39, 46, 54, 71, 69, 51, 69, 18, 22, 65, 75, 47, 48, 24, 51, 23, 62, 28, 56, 38, 47, 78, 27, 40, 20, 69, 46, 68, 47, 62, 71, 64, 77, 61, 53
