+++ train.html
@@ -1,0 +1,22 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="UTF-8">
+        <link type="text/css" rel="stylesheet" href="styles.css">
+        <title>Entrenamiento de K-Means</title>
+    </head>
+    <header style="background-color: lightblue;height: 5vh; display: flex; align-items: center;">
+        <ul style="margin: 0; list-style-type: none; font-weight: bold; padding-left: 20px;">
+            <li><a href="mainPage" style="text-decoration: none;color: black;">Pagina Principal</a></li>
+        </ul>
+    </header>
+    <body style="margin: 0;">
+        <div style="display: flex; justify-content: center;">
+            <h1>¡K-Means entrenado!</h1>
+        </div>
+        <div style="display: flex; justify-content: center;">
+            <button style="border-radius: 20px; font-size: 16px; margin: 0 10px;"><a href="getCentroids" style="text-decoration: none;color: black; font-weight: bold;">¡Obten tus centroides!</a></button>
+            <button style="border-radius: 20px; font-size: 16px; margin: 0 10px;"><a href="getLabels" style="text-decoration: none;color: black; font-weight: bold;">¡Obten tus etiquetas!</a></button>
+        </div>
+    </body>
+</html>
